requirements.txt
paho-mqtt==1.6.1
aiorun==2023.7.2
pymongo==4.5.0
netifaces==0.11.0
flask==2.0.3
werkzeug==2.0.3
jinja2==3.0.3
itsdangerous==2.0.1
flask-socketio==5.1.1
python-engineio==4.3.1
python-socketio==5.5.2
eventlet==0.33.3
python-dotenv==1.0.0 